Rearranged functions and added more functionality
