+++ src/main/java/dk/dtu/game/core/solver/AlgorithmX/ColumnNode.java
@@ -1,4 +1,4 @@
-package dk.dtu.game.solver;
+package dk.dtu.game.core.solver.AlgorithmX;
 
 public class ColumnNode extends Node {
     public int size;
