+++ src/main/java/dk/dtu/game/core/solver/AlgorithmX/Placement.java
@@ -1,0 +1,14 @@
+package dk.dtu.game.core.solver.AlgorithmX;
+
+public class Placement {
+    public int row;
+    public int col;
+    public int value;
+
+    public Placement(int row, int col, int value) {
+        this.row = row;
+        this.col = col;
+        this.value = value;
+    }
+
+}
