--- src/main/java/dk/dtu/game/solver/Placement.java
@@ -1,14 +1,0 @@
-package dk.dtu.game.solver;
-
-class Placement {
-    int row;
-    int col;
-    int value;
-
-    Placement(int row, int col, int value) {
-        this.row = row;
-        this.col = col;
-        this.value = value;
-    }
-
-}
