+++ src/test/java/dk/dtu/core/CreateTest.java
@@ -1,7 +1,7 @@
 package dk.dtu.core;
 
 import dk.dtu.game.core.Board;
-import dk.dtu.game.solver.solverAlgorithm;
+import dk.dtu.game.core.solver.BruteForce.BruteForceAlgorithm;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -55,7 +55,7 @@ class CreateTest {
         Board board = null;
         try {
             board = new Board(3, 3);
-            solverAlgorithm.fillBoard(board);
+            BruteForceAlgorithm.fillBoard(board);
 
         } catch (Exception e) {
             fail("Creating or filling the board should not throw an exception.");
@@ -128,7 +128,7 @@ class CreateTest {
     @DisplayName("Test solving a small 4x4 Sudoku board")
     void testSmallBoardSolving() throws Exception {
         Board board = new Board(2, 2);
-        solverAlgorithm.fillBoard(board);
+        BruteForceAlgorithm.fillBoard(board);
         assertTrue(isValidSudoku(board), "Small board should adhere to Sudoku rules.");
     }
 
@@ -138,8 +138,8 @@ class CreateTest {
         Board board = new Board(4, 4);
         assertNotNull(board, "Large board should be created successfully.");
         //Fill the board with a correct Sudoku solution first
-        solverAlgorithm.fillBoard(board);
-        assertTrue(solverAlgorithm.isValidSudoku(board.getBoard()), "Large board should adhere to Sudoku rules.");
+        BruteForceAlgorithm.fillBoard(board);
+        assertTrue(BruteForceAlgorithm.isValidSudoku(board.getBoard()), "Large board should adhere to Sudoku rules.");
         assertEquals(16, board.getDimensions(), "Board should have correct dimensions for an 16x16 board.");
     }
 
@@ -157,7 +157,7 @@ class CreateTest {
     void testValidSudokuValidation() throws Exception {
         // Fill the board with a correct Sudoku solution first
         Board board = new Board(3, 3);
-        solverAlgorithm.fillBoard(board);
+        BruteForceAlgorithm.fillBoard(board);
         assertTrue(isValidSudoku(board), "Board validation should confirm correct Sudoku.");
     }
 
@@ -167,7 +167,7 @@ class CreateTest {
         Board board = new Board(3, 3);
         board.setNumber(0, 0, 1);
         board.setNumber(0, 1, 1); // Create a conflict
-        assertFalse(solverAlgorithm.sudoku(board.getBoard()), "Solver should detect the board is unsolvable.");
+        assertFalse(BruteForceAlgorithm.sudoku(board.getBoard()), "Solver should detect the board is unsolvable.");
     }
 
     @Test
@@ -175,7 +175,7 @@ class CreateTest {
     void testSolverPerformance() throws Exception {
         Board board = new Board(3, 3);
         long startTime = System.currentTimeMillis();
-        solverAlgorithm.fillBoard(board);
+        BruteForceAlgorithm.fillBoard(board);
         long endTime = System.currentTimeMillis();
         assertTrue((endTime - startTime) < 1000, "Solver should complete within reasonable time for a standard board.");
     }
@@ -184,8 +184,8 @@ class CreateTest {
     @DisplayName("Create sudoku, solve it and test if valid")
     void testCreateSudoku() throws Exception {
         Board board = new Board(3, 3);
-        solverAlgorithm.createSudoku(board);
-        solverAlgorithm.sudoku(board.getBoard());
+        BruteForceAlgorithm.createSudoku(board);
+        BruteForceAlgorithm.sudoku(board.getBoard());
         assertTrue(isValidSudoku(board), "Board should adhere to Sudoku rules.");
     }
 
@@ -193,14 +193,14 @@ class CreateTest {
     @DisplayName("Add wrong numbers and test if sudoku solves correctly")
     void testInvalidSudoku() throws Exception {
         Board board = new Board(3, 3);
-        solverAlgorithm.createSudoku(board);
+        BruteForceAlgorithm.createSudoku(board);
         board.setInitialBoard(board.getBoard());
-        int [][] solvedBoard = solverAlgorithm.getSolutionBoard(board.getInitialBoard());
+        int [][] solvedBoard = BruteForceAlgorithm.getSolutionBoard(board.getInitialBoard());
 
         board.setNumber(0, 0, 1);
         board.setNumber(0, 1, 1); // Create a conflict
 
-        board.setBoard(solverAlgorithm.getSolutionBoard(board.getInitialBoard()));
+        board.setBoard(BruteForceAlgorithm.getSolutionBoard(board.getInitialBoard()));
 
         assertArrayEquals(solvedBoard, board.getBoard(), "The board should be the same as the solved board");
