+++ src/test/java/dk/dtu/core/UndoTest.java
@@ -3,7 +3,8 @@ package dk.dtu.core;
 import dk.dtu.engine.core.WindowManager;
 import dk.dtu.game.core.Move;
 import dk.dtu.game.core.SudokuGame;
-import dk.dtu.game.solver.solverAlgorithm;
+import dk.dtu.game.core.solver.BruteForce.BruteForceAlgorithm;
+import dk.dtu.game.core.solver.solverAlgorithm;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -21,7 +22,7 @@ public class UndoTest {
         // Manually initialize the board component to avoid NullPointerException
         game.createBoard(3, 3, 50); // Assuming this is the correct way to initialize the board in your game
 
-        solverAlgorithm.createSudoku(game.gameboard);
+        BruteForceAlgorithm.createSudoku(game.gameboard);
 
         // No need to call displayNumbersVisually here if it's just for visual representation and not part of the test logic
 
@@ -48,7 +49,7 @@ public class UndoTest {
         game.createBoard(3, 3, 50); // Set up the board
 
         // Generate a solvable Sudoku puzzle
-        solverAlgorithm.createSudoku(game.gameboard);
+        BruteForceAlgorithm.createSudoku(game.gameboard);
 
         // Remove numbers to generate hints (if not already part of createSudoku)
         // Assuming fillHintList() populates the hintList based on the current board state
@@ -60,7 +61,7 @@ public class UndoTest {
         }
 
         // Validate the Sudoku board
-        assertTrue(solverAlgorithm.isValidSudoku(game.gameboard.getBoard()), "Applying all hints should result in a valid Sudoku.");
+        assertTrue(BruteForceAlgorithm.isValidSudoku(game.gameboard.getBoard()), "Applying all hints should result in a valid Sudoku.");
     }
 
 }
